requirements.txt
streamlit
openai
pillow
pyngrok
